Hotfix migraciones

+++ database/migrations/2018_05_16_123611_agrega_campo_id_persona_creacion_a_actividad.php
@@ -27,8 +27,10 @@ class AgregaCampoIdPersonaCreacionAActividad extends Migration
      */
     public function down()
     {
-        Schema::table('Actividad', function (Blueprint $table) {
-            $table->dropColumn('idPersonaCreacion');
-        });
+        if (Schema::hasColumn('Actividad', 'idPersonaCreacion')) {
+            Schema::table('Actividad', function (Blueprint $table) {
+                $table->dropColumn('idPersonaCreacion');
+            });
+        }
     }
 }
